Trait libp2p_quic::Provider

pub trait Provider: Unpin + Send + Sized + 'static {
    type IfWatcher: Unpin + Send;

    // Required methods
    fn runtime() -> Runtime;
    fn new_if_watcher() -> Result<Self::IfWatcher>;
    fn poll_if_event(
        watcher: &mut Self::IfWatcher,
        cx: &mut Context<'_>
    ) -> Poll<Result<IfEvent>>;
    fn sleep(duration: Duration) -> BoxFuture<'static, ()>;
    fn send_to<'a>(
        udp_socket: &'a UdpSocket,
        buf: &'a [u8],
        target: SocketAddr
    ) -> BoxFuture<'a, Result<usize>>;
}

Provider for a corresponding quinn runtime and spawning tasks.

Required Associated Types

type IfWatcher: Unpin + Send

Required Methods

fn runtime() -> Runtime

Run the corresponding runtime

fn new_if_watcher() -> Result<Self::IfWatcher>

Create a new if_watch watcher that reports IfEvents for network interface changes.

fn poll_if_event( watcher: &mut Self::IfWatcher, cx: &mut Context<'_> ) -> Poll<Result<IfEvent>>

Poll for an address change event.

fn sleep(duration: Duration) -> BoxFuture<'static, ()>

Sleep for specified amount of time.

fn send_to<'a>( udp_socket: &'a UdpSocket, buf: &'a [u8], target: SocketAddr ) -> BoxFuture<'a, Result<usize>>

Sends data on the socket to the given address. On success, returns the number of bytes written.

Object Safety

This trait is not object safe.